requirements.txt
asgiref==3.4.1
attrs==21.2.0
coverage==5.5
Django==3.2.7
django-cleanup==5.2.0
djangorestframework==3.12.4
iniconfig==1.1.1
packaging==21.0
Pillow==9.2.0 
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
pytz==2021.1
sqlparse==0.4.2
toml==0.10.2
drf-yasg
